Add a script comparing the unicode version map to an older version
This script is used as `npm run compare-maps path/to/old/map.json`. It
will report any missing name (names which were in the old map but are not
available anymore) as well as any version change for a name. It will exit
with a failure exit code in case of missing names.